// src/kaspa/kaspa.d.ts
        interface UtxoProcessor {
            /**
            * @param {UtxoProcessorNotificationCallback} callback
            */
            addEventListener(callback: UtxoProcessorNotificationCallback): void;
            /**
            * @param {UtxoProcessorEventType} event
            * @param {UtxoProcessorNotificationCallback} [callback]
            */
            addEventListener<E extends keyof UtxoProcessorEventMap>(
                event: E,
                callback: UtxoProcessorNotificationCallback<E>
            )
        }


    /**
     * UtxoProcessor constructor arguments.
     * 
     * @see {@link UtxoProcessor}, {@link UtxoContext}, {@link RpcClient}, {@link NetworkId}
     * @category Wallet SDK
     */
    export interface IUtxoProcessorArgs {
        /**
         * The RPC client to use for network communication.
         */
        rpc : RpcClient;
        networkId : NetworkId | string;
    }
// ...
    export interface IUtxoIndexNotEnabledEvent {
        url? : string;
    }
    


    /**
     * Emitted by {@link UtxoProcessor} when it disconnects from RPC.
     * 
     * @category Wallet Events
     */
    export interface IDisconnectEvent {
        networkId : string;
        url? : string;
    }
    


    /**
     * Emitted by {@link UtxoProcessor} when it negotiates a successful RPC connection.
     * 
     * @category Wallet Events
     */
    export interface IConnectEvent {
        networkId : string;
        url? : string;
    }
// ...
        export enum UtxoProcessorEventType {
            Connect = "connect",
            Disconnect = "disconnect",
            UtxoIndexNotEnabled = "utxo-index-not-enabled",
            SyncState = "sync-state",
            UtxoProcStart = "utxo-proc-start",
            UtxoProcStop = "utxo-proc-stop",
            UtxoProcError = "utxo-proc-error",
            DaaScoreChange = "daa-score-change",
            Pending = "pending",
            Reorg = "reorg",
            Stasis = "stasis",
            Maturity = "maturity",
            Discovery = "discovery",
            Balance = "balance",
            Error = "error",
        }

        /**
         * {@link UtxoProcessor} notification event data map.
         * 
         * @category Wallet API
         */
        export type UtxoProcessorEventMap = {
            "connect": IConnectEvent,
            "disconnect": IDisconnectEvent,
            "utxo-index-not-enabled": IUtxoIndexNotEnabledEvent,
            "sync-state": ISyncStateEvent,
            "server-status": IServerStatusEvent,
            "utxo-proc-start": undefined,
            "utxo-proc-stop": undefined,
            "utxo-proc-error": IUtxoProcErrorEvent,
            "daa-score-change": IDaaScoreChangeEvent,
            "pending": IPendingEvent,
            "reorg": IReorgEvent,
            "stasis": IStasisEvent,
            "maturity": IMaturityEvent,
            "discovery": IDiscoveryEvent,
            "balance": IBalanceEvent,
            "error": IErrorEvent
        }

        /**
         * 
         * @category Wallet API
         */

        export type UtxoProcessorEvent<T extends keyof UtxoProcessorEventMap> = {
          [K in T]: {
            type: K,
            data: UtxoProcessorEventMap[K]
          }
        }[T];
        
        /**
         * {@link UtxoProcessor} notification callback type.
         * 
         * This type declares the callback function that is called when notification is emitted
         * from the UtxoProcessor or UtxoContext subsystems.
         * 
         * @see {@link UtxoProcessor}, {@link UtxoContext},
         * 
         * @category Wallet SDK
         */

        export type UtxoProcessorNotificationCallback<E extends keyof UtxoProcessorEventMap = keyof UtxoProcessorEventMap> = (event: UtxoProcessorEvent<E>) => void;
// ...
    export interface IUtxoContextArgs {
        /**
         * Associated UtxoProcessor.
         */
        processor: UtxoProcessor;
        /**
         * Optional id for the UtxoContext.
         * **The id must be a valid 32-byte hex string.**
         * You can use {@link sha256FromBinary} or {@link sha256FromText} to generate a valid id.
         * 
         * If not provided, a random id will be generated.
         * The IDs are deterministic, based on the order UtxoContexts are created.
         */
        id?: HexString;
    }
// ...
        interface RpcClient {
            /**
            * @param {RpcEventCallback} callback
            */
            addEventListener(callback:RpcEventCallback): void;
            /**
            * @param {RpcEventType} event
            * @param {RpcEventCallback} [callback]
            */
            addEventListener<M extends keyof RpcEventMap>(
                event: M,
                callback: (eventData: RpcEventMap[M]) => void
            )
        }
// ...
    export interface INewBlockTemplate {
        [key: string]: any;
    }
    


    /**
     * Pruning point UTXO set override notification event is produced when the
     * UTXO set override for the pruning point changes in the Kaspa BlockDAG.
     * 
     * @category Node RPC
     */
    export interface IPruningPointUtxoSetOverride {
        [key: string]: any;
    }
    


    /**
     * Virtual DAA score changed notification event is produced when the virtual
     * Difficulty Adjustment Algorithm (DAA) score changes in the Kaspa BlockDAG.
     * 
     * @category Node RPC
     */
    export interface IVirtualDaaScoreChanged {
        [key: string]: any;
    }
    


    /**
     * Sink blue score changed notification event is produced when the blue
     * score of the sink block changes in the Kaspa BlockDAG.
     * 
     * @category Node RPC
     */
    export interface ISinkBlueScoreChanged {
        [key: string]: any;
    }
    


    /**
     * UTXOs changed notification event is produced when the set
     * of unspent transaction outputs (UTXOs) changes in the
     * Kaspa BlockDAG. The event notification is scoped to the
     * monitored list of addresses specified during the subscription.
     * 
     * @category Node RPC
     */
    export interface IUtxosChanged {
        [key: string]: any;
    }
    


    /**
     * Finality conflict resolved notification event is produced when a finality
     * conflict in the Kaspa BlockDAG is resolved.
     * 
     * @category Node RPC
     */
    export interface IFinalityConflictResolved {
        [key: string]: any;
    }
    


    /**
     * Finality conflict notification event is produced when a finality
     * conflict occurs in the Kaspa BlockDAG.
     * 
     * @category Node RPC
     */
    export interface IFinalityConflict {
        [key: string]: any;
    }
    


    /**
     * Virtual chain changed notification event is produced when the virtual
     * chain changes in the Kaspa BlockDAG.
     * 
     * @category Node RPC
     */
    export interface IVirtualChainChanged {
        [key: string]: any;
    }
    


    /**
     * Block added notification event is produced when a new
     * block is added to the Kaspa BlockDAG.
     * 
     * @category Node RPC
     */
    export interface IBlockAdded {
        [key: string]: any;
    }
// ...
export type RpcEventMap = {
    "connect" : undefined,
    "disconnect" : undefined,
    "block-added" : IBlockAdded,
    "virtual-chain-changed" : IVirtualChainChanged,
    "finality-conflict" : IFinalityConflict,
    "finality-conflict-resolved" : IFinalityConflictResolved,
    "utxos-changed" : IUtxosChanged,
    "sink-blue-score-changed" : ISinkBlueScoreChanged,
    "virtual-daa-score-changed" : IVirtualDaaScoreChanged,
    "pruning-point-utxo-set-override" : IPruningPointUtxoSetOverride,
    "new-block-template" : INewBlockTemplate,
}

/**
 * RPC notification event.
 * 
 * @category Node RPC
 */
export type RpcEvent = {
    [K in keyof RpcEventMap]: { event: K, data: RpcEventMap[K] }
}[keyof RpcEventMap];
// ...
export type RpcEventCallback = (event: RpcEvent) => void;
// ...
export class Address {
/**
** Return copy of self without private attributes.
*/
  toJSON(): Object;
/**
* Return stringified version of self.
*/
  toString(): string;
  free(): void;
/**
* @param {string} address
*/
  constructor(address: string);
/**
* @param {string} address
* @returns {boolean}
*/
  static validate(address: string): boolean;
/**
* Convert an address to a string.
* @returns {string}
*/
  toString(): string;
/**
* @param {number} n
* @returns {string}
*/
  short(n: number): string;
/**
*/
  readonly payload: string;
/**
*/
  readonly prefix: string;
/**
*/
  setPrefix: string;
/**
*/
  readonly version: string;
}
// ...
export class Balance {
  free(): void;
/**
* @param {NetworkType | NetworkId | string} network_type
* @returns {BalanceStrings}
*/
  toBalanceStrings(network_type: NetworkType | NetworkId | string): BalanceStrings;
/**
* Confirmed amount of funds available for spending.
*/
  readonly mature: bigint;
/**
* Amount of funds that are being send and are not yet accepted by the network.
*/
  readonly outgoing: bigint;
/**
* Amount of funds that are being received and are not yet confirmed.
*/
  readonly pending: bigint;
}
/**
*
* Formatted string representation of the {@link Balance}.
*
* The value is formatted as `123,456.789`.
*
* @category Wallet SDK
*/
export class BalanceStrings {
  free(): void;
/**
*/
  readonly mature: string;
/**
*/
  readonly pending: string | undefined;
}
// ...
export class NetworkId {
/**
** Return copy of self without private attributes.
*/
  toJSON(): Object;
/**
* Return stringified version of self.
*/
  toString(): string;
  free(): void;
/**
* @param {any} value
*/
  constructor(value: any);
/**
* @returns {string}
*/
  toString(): string;
/**
* @returns {string}
*/
  addressPrefix(): string;
/**
*/
  readonly id: string;
/**
*/
  suffix?: number;
/**
*/
  type: NetworkType;
}
// ...
export class UtxoContext {
/**
** Return copy of self without private attributes.
*/
  toJSON(): Object;
/**
* Return stringified version of self.
*/
  toString(): string;
  free(): void;
/**
* @param {IUtxoContextArgs} js_value
*/
  constructor(js_value: IUtxoContextArgs);
/**
* Performs a scan of the given addresses and registers them in the context for event notifications.
* @param {(Address | string)[]} addresses
* @param {bigint | undefined} [optional_current_daa_score]
* @returns {Promise<void>}
*/
  trackAddresses(addresses: (Address | string)[], optional_current_daa_score?: bigint): Promise<void>;
/**
* Unregister a list of addresses from the context. This will stop tracking of these addresses.
* @param {(Address | string)[]} addresses
* @returns {Promise<void>}
*/
  unregisterAddresses(addresses: (Address | string)[]): Promise<void>;
/**
* Clear the UtxoContext.  Unregister all addresses and clear all UTXO entries.
* IMPORTANT: This function must be manually called when disconnecting or re-connecting to the node
* (followed by address re-registration).  
* @returns {Promise<void>}
*/
  clear(): Promise<void>;
/**
*
* Returns a range of mature UTXO entries that are currently
* managed by the UtxoContext and are available for spending.
*
* NOTE: This function is provided for informational purposes only.
* **You should not manage UTXO entries manually if they are owned by UtxoContext.**
*
* The resulting range may be less than requested if UTXO entries
* have been spent asynchronously by UtxoContext or by other means
* (i.e. UtxoContext has received notification from the network that
* UtxoEntries have been spent externally).
*
* UtxoEntries are kept in in the ascending sorted order by their amount.
* @param {number} from
* @param {number} to
* @returns {UtxoEntryReference[]}
*/
  getMatureRange(from: number, to: number): UtxoEntryReference[];
/**
* Returns pending UTXO entries that are currently managed by the UtxoContext.
* @returns {UtxoEntryReference[]}
*/
  getPending(): UtxoEntryReference[];
/**
* Current {@link Balance} of the UtxoContext.
*/
  readonly balance: Balance | undefined;
/**
* Current {@link BalanceStrings} of the UtxoContext.
*/
  readonly balanceStrings: BalanceStrings | undefined;
/**
*/
  readonly isActive: boolean;
/**
* Obtain the length of the mature UTXO entries that are currently
* managed by the UtxoContext.
*/
  readonly matureLength: number;
}
// ...
export class UtxoEntryReference {
/**
** Return copy of self without private attributes.
*/
  toJSON(): Object;
/**
* Return stringified version of self.
*/
  toString(): string;
  free(): void;
/**
* @returns {string}
*/
  toString(): string;
/**
*/
  readonly address: Address | undefined;
/**
*/
  readonly amount: bigint;
/**
*/
  readonly blockDaaScore: bigint;
/**
*/
  readonly entry: UtxoEntry;
/**
*/
  readonly isCoinbase: boolean;
/**
*/
  readonly outpoint: TransactionOutpoint;
/**
*/
  readonly scriptPublicKey: ScriptPublicKey;
}
// ...
export class UtxoProcessor {
/**
** Return copy of self without private attributes.
*/
  toJSON(): Object;
/**
* Return stringified version of self.
*/
  toString(): string;
  free(): void;
/**
* @param {UtxoProcessorEventType | UtxoProcessorEventType[] | string | string[]} event
* @param {UtxoProcessorNotificationCallback | undefined} [callback]
*/
  removeEventListener(event: UtxoProcessorEventType | UtxoProcessorEventType[] | string | string[], callback?: UtxoProcessorNotificationCallback): void;
/**
* UtxoProcessor constructor.
*
*
*
* @see {@link IUtxoProcessorArgs}
* @param {IUtxoProcessorArgs} js_value
*/
  constructor(js_value: IUtxoProcessorArgs);
/**
* Starts the UtxoProcessor and begins processing UTXO and other notifications.
* @returns {Promise<void>}
*/
  start(): Promise<void>;
/**
* Stops the UtxoProcessor and ends processing UTXO and other notifications.
* @returns {Promise<void>}
*/
  stop(): Promise<void>;
/**
* @param {NetworkId | string} network_id
*/
  setNetworkId(network_id: NetworkId | string): void;
/**
*
* Set the coinbase transaction maturity period DAA score for a given network.
* This controls the DAA period after which the user transactions are considered mature
* and the wallet subsystem emits the transaction maturity event.
*
* @see {@link TransactionRecord}
* @see {@link IUtxoProcessorEvent}
*
* @category Wallet SDK
* @param {NetworkId | string} network_id
* @param {bigint} value
*/
  static setCoinbaseTransactionMaturityDAA(network_id: NetworkId | string, value: bigint): void;
/**
*
* Set the user transaction maturity period DAA score for a given network.
* This controls the DAA period after which the user transactions are considered mature
* and the wallet subsystem emits the transaction maturity event.
*
* @see {@link TransactionRecord}
* @see {@link IUtxoProcessorEvent}
*
* @category Wallet SDK
* @param {NetworkId | string} network_id
* @param {bigint} value
*/
  static setUserTransactionMaturityDAA(network_id: NetworkId | string, value: bigint): void;
/**
*/
  readonly isActive: boolean;
/**
*/
  readonly networkId: string | undefined;
/**
*/
  readonly rpc: RpcClient;
}